Implement PHP7 features in API objects
- type declarations
- strict type declaration
- other minor non-functional changes

Content relation class and interface has not been touched, as they will
be refactored anyway.

This commit includes few changes in classes outside API folder where
casting was necessary on API objects method calls

+++ src/functions.php
@@ -163,7 +163,7 @@ function debug( $message ) {
  */
 function get_available_language_names( $related = true, $include_current_site = true ) {
 
-	$current_site_id = get_current_blog_id();
+	$current_site_id = (int) get_current_blog_id();
 
 	$related_sites = [];
